requirements.txt
django
django_cleanup
django-multiselectfield
django-summernote
